+++ Engine/src/utils/datafile.h
@@ -11,7 +11,7 @@ namespace prev {
 
 		inline void * GetData() const { return (void *)m_Data.c_str(); }
 		inline std::string GetDataStr() const { return m_Data; }
-		inline unsigned int GetNumBytes() const { return m_Data.size(); }
+		inline unsigned int GetNumBytes() const { return (unsigned int)m_Data.size(); }
 		inline std::string GetExtension() const { return m_Ext; }
 	private:
 		void Release();
